# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: support-management
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "12.4"
servers:
  - url: http://localhost:51578
    description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/errands/{errandId}/attachments:
    get:
      tags:
        - Errand attachments
      summary: Read errand attachments
      description: Fetches the errand attachments that matches the provided errand
        id
      operationId: readErrandAttachments
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrandAttachment"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
        - Errand attachments
      summary: Create errand attachment
      description: Creates a new errand attachment based on the supplied attributes
      operationId: createErrandAttachment
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                errandAttachment:
                  type: string
                  format: binary
              required:
                - errandAttachment
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}:
    get:
      tags:
        - Errands
      summary: Read errand
      description: Fetches the errand that matches the provided id
      operationId: readErrand
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errand"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - Errands
      summary: Delete errand
      description: Deletes the errand that matches the provided id
      operationId: deleteErrand
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
      responses:
        "204":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
        - Errands
      summary: Update errand
      description: Updates the errand matching provided id with the supplied attributes
      operationId: updateErrand
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Errand"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errand"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}/communication/{communicationId}/attachments/{attachmentId}:
    get:
      tags:
        - Errand communication
      summary: Get a streamed communication attachment.
      description: "Fetches the communication attachment that matches the provided\
        \ ID, in a streamed manner"
      operationId: getMessageAttachment
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
        - name: communicationId
          in: path
          description: communication ID
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
        - name: attachmentId
          in: path
          description: Message attachment ID
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
      responses:
        "200":
          description: Successful Operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "507":
          description: Insufficient Storage
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}/communication/conversations/{conversationId}/messages/{messageId}/attachments/{attachmentId}:
    get:
      tags:
        - Errand communication
      summary: Get a streamed conversation message attachment.
      description: "Fetches the conversation message attachment that matches the provided\
        \ ID, in a streamed manner"
      operationId: getConversationMessageAttachment
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
        - name: conversationId
          in: path
          description: Conversation ID
          required: true
          schema:
            type: string
          example: 1aefbbb8-de82-414b-b5d7-ba7c5bbe4506
        - name: messageId
          in: path
          description: Message ID
          required: true
          schema:
            type: string
          example: be3a5c7b-0f39-4867-8065-6a286b36fcc3
        - name: attachmentId
          in: path
          description: Conversation message attachment ID
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
      responses:
        "200":
          description: Successful Operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "507":
          description: Insufficient Storage
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/{namespace}/errands/{errandId}/attachments/{attachmentId}:
    get:
      tags:
        - Errand attachments
      summary: Read errand attachment
      description: Fetches the errand attachment that matches the provided errand
        id and attachment id
      operationId: readErrandAttachment
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
        - name: attachmentId
          in: path
          description: Errand attachment id
          required: true
          schema:
            type: string
          example: 5f79a808-0ef3-4985-99b9-b12f23e202a7
      responses:
        "200":
          description: Successful Operation
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - Errand attachments
      summary: Delete errand attachment
      description: Deletes the errand attachment that matches the provided errand
        id and attachment id
      operationId: deleteErrandAttachment
      parameters:
        - name: namespace
          in: path
          description: Namespace
          required: true
          schema:
            type: string
            pattern: "[\\w|\\-]+"
          example: MY_NAMESPACE
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: errandId
          in: path
          description: Errand id
          required: true
          schema:
            type: string
          example: b82bd8ac-1507-4d9a-958d-369261eecc15
        - name: attachmentId
          in: path
          description: Errand attachment id
          required: true
          schema:
            type: string
          example: 5f79a808-0ef3-4985-99b9-b12f23e202a7
      responses:
        "204":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: { }
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Errand:
      type: object
      description: Errand model
      properties:
        id:
          type: string
          description: Unique id for the errand
          examples:
            - f0882f1d-06bc-47fd-b017-1d8307f5ce95
          readOnly: true
        errandNumber:
          type: string
          description: Unique number for the errand
          examples:
            - KC-23010001
          readOnly: true
        title:
          type: string
          description: Title for the errand
          examples:
            - Title for the errand
        priority:
          $ref: "#/components/schemas/Priority"
        stakeholders:
          type: array
          items:
            $ref: "#/components/schemas/Stakeholder"
        externalTags:
          type: array
          items:
            $ref: "#/components/schemas/ExternalTag"
          uniqueItems: true
        parameters:
          type: array
          description: Parameters for the errand
          items:
            $ref: "#/components/schemas/Parameter"
        jsonParameters:
          type: array
          description: JSON parameters for the errand
          items:
            $ref: "#/components/schemas/JsonParameter"
        classification:
          $ref: "#/components/schemas/Classification"
        status:
          type: string
          description: Status for the errand
          examples:
            - NEW_CASE
        resolution:
          type: string
          description: Resolution status for closed errands. Value can be set to anything
          examples:
            - FIXED
        description:
          type: string
          description: Errand description text
          examples:
            - Order cake for everyone
        channel:
          type: string
          description: The channel from which the errand originated
          examples:
            - THE_CHANNEL
          maxLength: 255
        reporterUserId:
          type: string
          description: User id for the person which has created the errand
          examples:
            - joe01doe
        assignedUserId:
          type: string
          description: Id for the user which currently is assigned to the errand if
            a user is assigned
          examples:
            - joe01doe
        assignedGroupId:
          type: string
          description: Id for the group which is currently assigned to the errand
            if a group is assigned
          examples:
            - hardware support
        escalationEmail:
          type: string
          format: email
          description: Email address used for escalation of errand
          examples:
            - joe.doe@email.com
        contactReason:
          type: string
          description: Contact reason for the errand
          examples:
            - The printer is not working
        contactReasonDescription:
          type: string
          description: Contact reason description for the errand
          examples:
            - The printer is not working since the power cord is missing
          maxLength: 4096
        suspension:
          $ref: "#/components/schemas/Suspension"
          description: Suspension information
        businessRelated:
          type: boolean
          description: Flag to indicate if the errand is business related
          examples:
            - "true"
        labels:
          type: array
          description: List of labels for the errand
          items:
            $ref: "#/components/schemas/ErrandLabel"
        activeNotifications:
          type: array
          description: List of active notifications for the errand
          items:
            $ref: "#/components/schemas/Notification"
          readOnly: true
        created:
          type: string
          format: date-time
          description: Timestamp when errand was created
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        modified:
          type: string
          format: date-time
          description: Timestamp when errand was last modified
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
        touched:
          type: string
          format: date-time
          description: Timestamp when errand was last touched (created or modified)
          examples:
            - 2000-10-31T01:30:00.000+02:00
          readOnly: true
...
    ExternalTag:
      type: object
      description: External tag model
      properties:
        key:
          type: string
          description: Key for external tag
          examples:
            - caseId
        value:
          type: string
          description: Value for external tag
          examples:
            - 8849-2848
...
    JsonParameter:
      type: object
      description: JSON Parameter model
      properties:
        key:
          type: string
          description: Parameter key/name
          examples:
            - formData1
          minLength: 1
        value:
          $ref: "#/components/schemas/JsonNode"
          description: JSON structure value
          example:
            firstName: Joe
            lastName: Doe
        schemaId:
          type: string
          description: ID referencing a schema in the json-schema service
          examples:
            - 2281_person_1.0
          minLength: 1
      required:
        - key
        - schemaId
        - value
...
    Parameter:
      type: object
      description: Parameter model
      properties:
        key:
          type: string
          description: Parameter key
          minLength: 1
        displayName:
          type: string
          description: Parameter display name
        group:
          type: string
          description: Parameter group name
        values:
          type: array
          description: Parameter values. Each value can have a maximum length of 2000
            characters
          items:
            type: string
            maxLength: 2000
            minLength: 0
      required:
        - key
    Priority:
      type: string
      description: Priority model
      enum:
        - LOW
        - MEDIUM
        - HIGH
...
    ErrandAttachment:
      type: object
      description: ErrandAttachment model
      properties:
        id:
          type: string
          description: Unique identifier for the attachment
          examples:
            - cb20c51f-fcf3-42c0-b613-de563634a8ec
          readOnly: true
        fileName:
          type: string
          description: Name of the file
          examples:
            - my-file.txt
        mimeType:
          type: string
          description: Mime type of the file
          readOnly: true
        created:
          type: string
          format: date-time
          description: The attachment created date
          examples:
            - 2023-01-01T00:00:00Z
...
  securitySchemes: { }